norminette en algunos archhivos y quitado la carpeta tokenizer y pasado el split a utils cmd_split
