+++ Makefile
@@ -35,7 +35,7 @@ SRCS 					:= src/main.c  \
 						src/lexer/lexer_list.c \
 						src/lexer/cleaner.c \
 						src/builtins/bt_init.c \
-						src/tokenize/tokenizer.c \
+						src/utils/cmd_split.c \
 						src/parser/parse_env.c 
 
 OBJS = $(patsubst $(SRC_DIR)%.c,$(OBJ_DIR)%.o,$(SRCS))
